+++ data-protocols/ids/ids-jsonld-serdes-lib/build.gradle.kts
@@ -1,0 +1,31 @@
+/*
+ *  Copyright (c) 2022 Fraunhofer Institute for Software and Systems Engineering
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
+ *
+ */
+
+plugins {
+    `java-library`
+    `maven-publish`
+}
+
+dependencies {
+    testImplementation(project(":data-protocols:ids:ids-core"))
+}
+
+publishing {
+    publications {
+        create<MavenPublication>("ids-jsonld-serdes-lib") {
+            artifactId = "ids-jsonld-serdes-lib"
+            from(components["java"])
+        }
+    }
+}
