+++ docs/developer/modules.md
@@ -21,7 +21,6 @@ org.eclipse.dataspaceconnector:http:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:http-provisioner:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:http-receiver:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:jdk-logger-monitor:0.0.1-SNAPSHOT
-org.eclipse.dataspaceconnector:jsonld-serdes:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:junit:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:sql:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:dataloader-cli:0.0.1-SNAPSHOT
@@ -42,6 +41,7 @@ org.eclipse.dataspaceconnector:ids-api-configuration:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-api-multipart-dispatcher-v1:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-api-multipart-endpoint-v1:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-core:0.0.1-SNAPSHOT
+org.eclipse.dataspaceconnector:ids-jsonld-serdes-lib:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-token-validation:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids-api-transform-v1:0.0.1-SNAPSHOT
