+++ settings.gradle.kts
@@ -40,6 +40,7 @@ include(":core:micrometer")
 include(":data-protocols:ids:ids-api-multipart-endpoint-v1")
 include(":data-protocols:ids:ids-api-multipart-dispatcher-v1")
 include(":data-protocols:ids:ids-core")
+include(":data-protocols:ids:ids-jsonld-serdes-lib")
 include(":data-protocols:ids:ids-spi")
 include(":data-protocols:ids:ids-transform-v1")
 include(":data-protocols:ids:ids-token-validation")
@@ -127,7 +128,6 @@ include(":extensions:data-plane:data-plane-api")
 include(":extensions:data-plane:integration-tests")
 include(":extensions:http-receiver")
 include(":extensions:http-provisioner")
-include(":extensions:jsonld-serdes")
 
 include(":extensions:sql:asset-index-sql")
 include(":extensions:sql:common-sql")
